ci: bound Wayland verifier protocol stalls
